feat(commerce): implement core commerce service features and feign clients
- Implement business logic for shopping cart, store, and warehouse services
- Define Feign clients for inter-service communication between commerce modules
- Refactor Cart entity to store products as a map instead of separate CartItem entities
- Add global exception handling and custom exceptions for warehouse operations
- Introduce pagination and sorting DTOs with corresponding mappers

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interaction/dto/cart/ChangeProductQuantityRequest.java
@@ -18,5 +18,5 @@ public class ChangeProductQuantityRequest {
 
     @NotNull(message = "Количество товара должно быть указано")
     @PositiveOrZero(message = "Количество товара не может быть отрицательным")
-    private Long newQuantity;
+    private Integer newQuantity;
 }
